scripts/bind_requests_xml: initial version
It misses a few features but the basics are working. It uses the XML
version of the Bind's statistics. The JSON would be simpler but the Bind
packages used by CentOS doesn't have JSON support and I didn't want to
create two separate versions of this script.

It uses xmlstarlet to gather data from the XML so you have to install
that first. It also uses curl to fetch the XML from Bind but you
probably have one already installed.